Refactored the [...slug] code
