+++ pages/[version]/[...slug].tsx
@@ -1,10 +1,4 @@
-import React, {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { useMemo } from "react";
 import {
   getArticleSlugFromString,
   getArticleSlugs,
@@ -15,27 +9,14 @@ import {
 import fs from "fs";
 import matter from "gray-matter";
 import { basename } from "path";
-import { useRouter } from "next/router";
 import Markdoc from "@markdoc/markdoc";
-import SideNav from "../../components/SideNav/SideNav";
-import TopNav from "../../components/TopNav/TopNav";
-import CategoriesNav from "../../components/CategoriesNav/CategoriesNav";
-import classNames from "classnames";
-import Footer from "../../components/Footer/Footer";
-import { components, configs } from "../../lib/markdoc";
-import Breadcrumb from "../../components/Breadcrumb/Breadcrumb";
+import { configs } from "../../lib/markdoc";
 import { MenuItem, PathObj } from "../../interface/common.interface";
-import { getCategoryByIndex } from "../../lib/utils";
 import ErrorBoundary from "../../components/ErrorBoundary";
-import { StepsContextProvider } from "../../context/StepsContext";
-import { has, isEmpty, startCase } from "lodash";
-import SkeletonLoader from "../../components/common/SkeletonLoader/SkeletonLoader";
-import { SKELETON_PARAGRAPH_WIDTHS } from "../../constants/SkeletonLoader.constants";
-import { useRouteChangingContext } from "../../context/RouteChangingContext";
 import Script from "next/script";
 import { SelectOption } from "../../components/SelectDropdown/SelectDropdown";
-import { useNavBarCollapsedContext } from "../../context/NavBarCollapseContext";
 import { getFormattedPartials } from "../../utils/CommonUtils";
+import DocsPageLayout from "../../components/PageLayouts/DocsPageLayout";
 
 interface Props {
   menu: MenuItem[];
@@ -45,10 +26,6 @@ interface Props {
   partials: Record<string, string>;
 }
 
-// Offset of 152px = 112px top nav-bar height + 240px top margin to show the link properly
-// as auto collapse of the sideNav may move the content
-const SCROLLING_OFFSET = 352;
-
 export default function Article({
   menu,
   content,
@@ -56,26 +33,6 @@ export default function Article({
   versionsList,
   partials,
 }: Props) {
-  const router = useRouter();
-  const { isRouteChanging } = useRouteChangingContext();
-  const { isMobileDevice } = useNavBarCollapsedContext();
-  const [sideNavCollapsed, setSideNavCollapsed] = useState<boolean>(false);
-
-  // Ref to keep track if the side nav is collapsed before, when "code-preview-container" is in the view.
-  const autoCollapsed = useRef(false);
-
-  const handleSideNavCollapsed = useCallback(
-    (value: boolean) => {
-      setSideNavCollapsed(value);
-    },
-    [setSideNavCollapsed]
-  );
-
-  const category = useMemo(
-    () => getCategoryByIndex(router.asPath, 2),
-    [router.asPath]
-  );
-
   const ast = useMemo(() => Markdoc.parse(content), [content]);
 
   const formattedPartialsObj = useMemo(
@@ -83,7 +40,7 @@ export default function Article({
     [partials]
   );
 
-  const ParsedContent = useMemo(
+  const parsedContent = useMemo(
     () =>
       Markdoc.transform(ast, {
         ...configs,
@@ -91,46 +48,6 @@ export default function Article({
       }),
     [ast, configs, formattedPartialsObj]
   );
-
-  const item = useMemo(
-    () =>
-      (menu as MenuItem[]).find(
-        (item) => getCategoryByIndex(item.url, 1) === category
-      ),
-    [menu, category]
-  );
-
-  useEffect(() => {
-    if (isMobileDevice) {
-      document.body.classList.add("min-width-600");
-    }
-  }, [isMobileDevice]);
-
-  // Function to scroll element into view with some offset margin
-  // For scrolling to the hash element on page after load
-  const scrollToElementWithOffsetMargin = useCallback(() => {
-    if (
-      has(window, "location.hash") &&
-      !isEmpty(window.location.hash) &&
-      !autoCollapsed.current // To prevent scrolling after each time the side panel collapses
-    ) {
-      const hashElementId = window.location.hash.slice(1);
-      const element = document.getElementById(hashElementId);
-
-      setTimeout(() => {
-        element &&
-          element.scrollIntoView({
-            block: "center",
-            inline: "center",
-            behavior: "smooth",
-          });
-      }, 500);
-    }
-  }, [autoCollapsed.current]);
-
-  useEffect(() => {
-    scrollToElementWithOffsetMargin();
-  });
 
   return (
     <>
@@ -150,51 +67,12 @@ export default function Article({
         }}
       />
       <ErrorBoundary>
-        <StepsContextProvider>
-          <div className="flex flex-col">
-            <div className="nav-bar-container">
-              <TopNav versionsList={versionsList} />
-              <CategoriesNav menu={menu} />
-            </div>
-            <div className="flex">
-              <SideNav
-                sideNavCollapsed={sideNavCollapsed}
-                category={item ? item.category : startCase(category)}
-                items={item ? item.children : []}
-                loading={isRouteChanging}
-                handleSideNavCollapsed={handleSideNavCollapsed}
-                ref={autoCollapsed}
-              />
-              <div
-                className={classNames(
-                  "content",
-                  sideNavCollapsed
-                    ? "collapsed-content"
-                    : "non-collapsed-content"
-                )}
-              >
-                <main className={classNames("flex flex-col mx-12 my-6")}>
-                  {isRouteChanging ? (
-                    <SkeletonLoader
-                      paragraph={{
-                        rows: SKELETON_PARAGRAPH_WIDTHS.length,
-                        width: SKELETON_PARAGRAPH_WIDTHS,
-                      }}
-                    />
-                  ) : (
-                    <>
-                      <Breadcrumb slug={slug} />
-                      {Markdoc.renderers.react(ParsedContent, React, {
-                        components,
-                      })}
-                    </>
-                  )}
-                </main>
-                <Footer />
-              </div>
-            </div>
-          </div>
-        </StepsContextProvider>
+        <DocsPageLayout
+          parsedContent={parsedContent}
+          menu={menu}
+          slug={slug}
+          versionsList={versionsList}
+        />
       </ErrorBoundary>
     </>
   );
